requirements.txt
openai
gpt_index==0.4.24
PyPDF2
PyCryptodome
gradio
urllib3==1.26.6
langchain==0.0.118
fastapi[all]
uvicorn
async-generator
async-exit-stack
cryptography
python-dotenv
beautifulsoup4
docx
EbookLib
numpy
odfpy
protobuf
psutil
pydub
python_docx
python_pptx
requests
scipy
striprtf==0.0.22
twilio
plotly
scikit-learn
termcolor
colorama